+++ src/olmo_core/distributed/checkpoint/filesystem.py
@@ -6,7 +6,13 @@ import os
 import pickle
 import tempfile
 import traceback
-from concurrent.futures import ThreadPoolExecutor, as_completed
+from concurrent.futures import (
+    Executor,
+    ProcessPoolExecutor,
+    ThreadPoolExecutor,
+    as_completed,
+)
+from contextlib import ExitStack
 from dataclasses import dataclass
 from functools import partial
 from pathlib import Path
@@ -15,6 +21,7 @@ from typing import Dict, List, Optional, Sequence, Tuple, cast
 import torch
 import torch.distributed as dist
 import torch.distributed.checkpoint as dist_cp
+import torch.multiprocessing as mp
 from bettermap import ordered_map_per_thread
 from torch.distributed.checkpoint.filesystem import WriteResult
 from torch.distributed.checkpoint.metadata import Metadata, MetadataIndex, StorageMeta
@@ -145,6 +152,22 @@ def _write_items(
     return results
 
 
+def _write_buckets(
+    buckets: List[List[WriteItem]], paths: List[str], planner: dist_cp.SavePlanner
+) -> List[WriteResult]:
+    results: List[WriteResult] = []
+    for bucket, path in zip(buckets, paths):
+        key = os.path.basename(path)
+        try:
+            results.extend(_write_items(path, key, bucket, planner))
+        except BaseException:
+            # NOTE: we might get an error here that can't be pickled, which causes a different failure
+            # later when PyTorch tries to reduce that error across ranks. So here we just make
+            # sure we're raising a simple error type that can be pickled.
+            raise OLMoCheckpointError(f"Original error:\n{traceback.format_exc()}")
+    return results
+
+
 def _narrow_tensor_by_index(
     tensor: torch.Tensor, offsets: Sequence[int], sizes: Sequence[int]
 ) -> torch.Tensor:
@@ -171,6 +194,7 @@ class RemoteFileSystemWriter(dist_cp.StorageWriter):
         self,
         path: PathOrStr,
         thread_count: Optional[int] = None,
+        process_count: Optional[int] = None,
         process_group: Optional[dist.ProcessGroup] = None,
         throttle_uploads: bool = False,
     ) -> None:
@@ -179,6 +203,7 @@ class RemoteFileSystemWriter(dist_cp.StorageWriter):
             raise ValueError("thread count must be at least 1")
         self.path = normalize_path(path)
         self.thread_count = thread_count or get_default_thread_count()
+        self.process_count = process_count
         self.process_group = process_group
         self.throttle_uploads = throttle_uploads
         self.save_id = generate_uuid()
@@ -222,35 +247,35 @@ class RemoteFileSystemWriter(dist_cp.StorageWriter):
             file_count += 1
             return f"{self.path}/{file_name}"
 
-        def write_buckets(buckets: List[List[WriteItem]]) -> List[WriteResult]:
-            results: List[WriteResult] = []
-            for bucket in buckets:
-                path = gen_file_name()
-                key = os.path.basename(path)
-                try:
-                    results.extend(_write_items(path, key, bucket, planner))
-                except BaseException:
-                    # NOTE: we might get an error here that can't be pickled, which causes a different failure
-                    # later when PyTorch tries to reduce that error across ranks. So here we just make
-                    # sure we're raising a simple error type that can be pickled.
-                    raise OLMoCheckpointError(f"Original error:\n{traceback.format_exc()}")
-            return results
-
         results: List[WriteResult]
         if self.throttle_uploads and is_url(self.path):
             buckets = _split_by_size_and_type(1, plan.items)
+            paths = [gen_file_name() for _ in buckets]
             results = do_n_at_a_time(
-                partial(write_buckets, buckets),
+                partial(_write_buckets, buckets, paths, planner),
                 process_group=self.process_group,
                 n=max(get_num_nodes() // 4, 1),
             )
         else:
             buckets = _split_by_size_and_type(self.thread_count, plan.items)
+            paths = [gen_file_name() for _ in buckets]
             results = []
-            with ThreadPoolExecutor(max_workers=self.thread_count) as executor:
+            with ExitStack() as stack:
+                executor: Executor
+                if self.process_count is not None and self.process_count > 1:
+                    executor = stack.enter_context(
+                        ProcessPoolExecutor(
+                            max_workers=self.process_count, mp_context=mp.get_context("spawn")
+                        )
+                    )
+                else:
+                    executor = stack.enter_context(
+                        ThreadPoolExecutor(max_workers=self.thread_count)
+                    )
+
                 futures = []
-                for bucket in buckets:
-                    futures.append(executor.submit(write_buckets, [bucket]))
+                for bucket, path in zip(buckets, paths):
+                    futures.append(executor.submit(_write_buckets, [bucket], [path], planner))
                 for f in as_completed(futures):
                     results.extend(f.result())