+++ src/test/distributed/checkpoint/filesystem_test.py
@@ -13,7 +13,9 @@ from olmo_core.testing import BACKENDS, run_distributed_test
 from olmo_core.utils import get_default_device
 
 
-def run_save_and_load_with_dtensors(dir, throttle: bool = False):
+def run_save_and_load_with_dtensors(
+    dir, thread_count: int | None = 2, process_count: int | None = None, throttle: bool = False
+):
     mesh = init_device_mesh(get_default_device().type, (dist.get_world_size(),))
 
     x_full = torch.randn(4, 4, device=get_default_device())
@@ -30,7 +32,12 @@ def run_save_and_load_with_dtensors(dir, throttle: bool = False):
     distcp.state_dict_saver.save(
         {"x": x, "y": y},
         checkpoint_id=dir,
-        storage_writer=RemoteFileSystemWriter(dir, thread_count=2, throttle_uploads=throttle),
+        storage_writer=RemoteFileSystemWriter(
+            dir,
+            thread_count=thread_count,
+            process_count=process_count,
+            throttle_uploads=throttle,
+        ),
     )
 
     # Now create new sharded copies with a different sharding strategy and load the checkpoint.
@@ -39,7 +46,7 @@ def run_save_and_load_with_dtensors(dir, throttle: bool = False):
     distcp.state_dict_loader.load(
         {"x": x_loaded, "y": y_loaded},
         checkpoint_id=dir,
-        storage_reader=RemoteFileSystemReader(dir, thread_count=2),
+        storage_reader=RemoteFileSystemReader(dir, thread_count=thread_count),
     )
 
     # Make sure the loaded tensors match the original tensors.
@@ -50,18 +57,26 @@ def run_save_and_load_with_dtensors(dir, throttle: bool = False):
 
 
 @pytest.mark.parametrize("backend", BACKENDS)
-def test_save_and_load_locally_with_dtensors(backend, tmp_path):
+@pytest.mark.parametrize(
+    "thread_count, process_count",
+    [pytest.param(2, None, id="threads"), pytest.param(None, 2, id="processes")],
+)
+def test_save_and_load_locally_with_dtensors(backend, tmp_path, thread_count, process_count):
     run_distributed_test(
         run_save_and_load_with_dtensors,
         backend=backend,
-        func_args=(tmp_path,),
+        func_args=(tmp_path, thread_count, process_count),
         start_method="spawn",
     )
 
 
 @pytest.mark.parametrize("backend", BACKENDS)
-@pytest.mark.parametrize("throttle", [True, False])
-def test_save_and_load_remotely_to_s3_with_dtensors(backend, s3_checkpoint_dir, throttle):
+@pytest.mark.parametrize(
+    "thread_count, process_count, throttle", [(2, None, True), (2, None, False)]
+)
+def test_save_and_load_remotely_to_s3_with_dtensors(
+    backend, s3_checkpoint_dir, thread_count, process_count, throttle
+):
     from botocore.exceptions import NoCredentialsError
 
     try:
@@ -72,7 +87,7 @@ def test_save_and_load_remotely_to_s3_with_dtensors(backend, s3_checkpoint_dir, 
     run_distributed_test(
         run_save_and_load_with_dtensors,
         backend=backend,
-        func_args=(s3_checkpoint_dir, throttle),
+        func_args=(s3_checkpoint_dir, thread_count, process_count, throttle),
         start_method="spawn",  # NOTE: forking causes a crash with boto3
     )